# backend/swagger.yaml
openapi: 3.0.3
info:
  title: MyActivities API
  description: |
    API REST de la plateforme **MyActivities** — découverte, planification et évaluation d'activités locales.

    ## Authentification
    L'API utilise des **JWT Bearer tokens** (access token 15 min + refresh token 7 jours).
    Les endpoints protégés nécessitent le header : `Authorization: Bearer <access_token>`

    ## Rôles utilisateurs
    | Rôle | Description |
    |------|-------------|
    | `visitor` | Non connecté — lecture seule |
    | `member` | Compte créé — planning, avis |
    | `organizer` | Référence ses activités |
    | `admin` | Modération et administration |

    ## Règles métier importantes
    - **1 avis par utilisateur par activité** (enforced côté backend)
    - **Toute nouvelle activité requiert une validation admin** avant publication
    - **Brute force** : 5 tentatives échouées → blocage 15 min
    - **Géolocalisation** : consentement explicite requis (RGPD Art. 6.1.a)
    - **Suppression de compte** : efface toutes les données personnelles (RGPD Art. 17)

    ## Codes d'erreur communs
    | Code | Signification |
    |------|---------------|
    | 400 | Données invalides |
    | 401 | Token manquant ou expiré |
    | 403 | Accès refusé (rôle insuffisant) |
    | 404 | Ressource introuvable |
    | 409 | Conflit (ex : email déjà utilisé) |
    | 429 | Trop de requêtes (rate limiting) |

  version: 1.0.0
  contact:
    name: Charles Deschamps
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT (expiration 15 min)

  # ─────────────────────────────────────────────
  # SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Le champ email est invalide.
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ── Auth ────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, pseudo, password]
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        pseudo:
          type: string
          minLength: 3
          maxLength: 30
          example: JohnDoe42
        password:
          type: string
          format: password
          minLength: 8
          description: "Min 8 caractères, 1 majuscule, 1 chiffre"
          example: MyP@ssw0rd
        siret:
          type: string
          pattern: '^[0-9]{14}$'
          description: >
            SIRET (14 chiffres) d'un organisateur. Fournir un SIRET crée
            directement un compte `organizer` ; l'omettre crée un compte `member`.
          example: "73282932000074"
    LoginRequest:
      type: object
      required: [pseudo, password]
      properties:
        pseudo:
          type: string
          minLength: 3
          maxLength: 30
          example: JohnDoe42
        password:
          type: string
          format: password
          example: MyP@ssw0rd
        rememberMe:
          type: boolean
          default: false
          description: Prolonge la durée du refresh token
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT valide 15 minutes
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Token opaque valide 7 jours
          example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        expiresIn:
          type: integer
          description: Durée de vie de l'access token en secondes
          example: 900
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          description: Token reçu par email (valide 1h, usage unique)
        password:
          type: string
          format: password
          minLength: 8
          example: NewP@ssw0rd

    # ── User ────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: john.doe@example.com
        pseudo:
          type: string
          example: JohnDoe42
        role:
          type: string
          enum: [member, organizer, admin]
          example: member
        status:
          type: string
          enum: [active, suspended, deleted]
          example: active
        siret:
          type: string
          nullable: true
          description: SIRET de l'organisateur (14 chiffres) ; null pour un membre.
          example: "73282932000074"
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://cdn.myactivities.app/avatars/uuid.jpg
        createdAt:
          type: string
          format: date-time
        emailVerified:
          type: boolean
          example: true
...
    Friend:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pseudo:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        friendsSince:
          type: string
          format: date-time
          nullable: true
    FriendRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        direction:
          type: string
          enum: [incoming, outgoing]
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            pseudo:
              type: string
            avatarUrl:
              type: string
              format: uri
              nullable: true
        createdAt:
          type: string
          format: date-time
...
    SendFriendRequestBody:
      type: object
      required: [addresseeId]
      properties:
        addresseeId:
          type: string
          format: uuid
...
  responses:
    Unauthorized:
      description: Token manquant ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Token JWT manquant ou expiré.
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Ressource introuvable.
    TooManyRequests:
      description: Trop de tentatives — rate limiting déclenché
      headers:
        Retry-After:
          schema:
            type: integer
          description: Secondes avant de réessayer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: RATE_LIMITED
            message: Compte temporairement bloqué. Réessayez dans 15 minutes.

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Créer un compte
      description: |
        Inscription avec email + pseudo + mot de passe.
        Un email de confirmation est envoyé — le compte n'est activé qu'après validation.
        Protection OWASP : le message d'erreur ne révèle pas si l'email existe déjà.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Compte créé — email de confirmation envoyé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Compte créé. Vérifiez votre email pour activer votre compte.
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email ou pseudo déjà utilisé (codes EMAIL_ALREADY_EXISTS / PSEUDO_ALREADY_EXISTS)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: EMAIL_ALREADY_EXISTS
                message: Un compte avec cet email existe déjà.
  /auth/login:
    post:
      tags: [Auth]
      summary: Se connecter
      description: |
        Connexion par pseudo + mot de passe.
        **Rate limiting** : 5 tentatives échouées → blocage 15 min (BNF-010).
        Retourne un access token (15 min) et un refresh token (7 jours).
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AuthTokens'
                  - type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_CREDENTIALS
                message: Pseudo ou mot de passe incorrect.
        '403':
          description: Compte suspendu (ACCOUNT_SUSPENDED) ou email non vérifié (EMAIL_NOT_VERIFIED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: EMAIL_NOT_VERIFIED
                message: Email non vérifié. Consultez votre boîte mail pour activer votre compte.
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Se déconnecter
      description: Invalide le refresh token côté serveur.
      operationId: logout
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '204':
          description: Déconnexion réussie
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir le token d'accès
      description: Échange un refresh token valide contre un nouvel access token.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Nouveau token émis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Refresh token invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Demande de réinitialisation de mot de passe
      description: |
        Envoie un email avec un lien de réinitialisation (valide 1h, usage unique).
        **Protection OWASP** : la réponse est identique que l'email existe ou non (anti-énumération).
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Email envoyé si le compte existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Si un compte correspond à cet email, vous recevrez un lien de réinitialisation.
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Réinitialiser le mot de passe
      description: Finalise la réinitialisation avec le token reçu par email.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Mot de passe réinitialisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Mot de passe mis à jour. Vous pouvez vous connecter.
        '400':
          description: Token invalide, expiré ou déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/verify-email/{token}:
    get:
      tags: [Auth]
      summary: Vérifier son adresse email
      description: Active le compte via le token reçu par email lors de l'inscription (valide 24h, usage unique).
      operationId: verifyEmail
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email vérifié
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email vérifié avec succès.
        '400':
          description: Token invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_VERIFICATION_TOKEN
                message: Token de vérification invalide ou expiré.

  # ══════════════════════════════
  # USERS
  # ══════════════════════════════
...
  /friends:
    get:
      tags: [Friends]
      summary: Lister ses amis
      operationId: listFriends
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des amis
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Friend'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /friends/requests:
    get:
      tags: [Friends]
      summary: Lister ses demandes en attente
      description: Demandes reçues (`incoming`) et envoyées (`outgoing`).
      operationId: listFriendRequests
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Demandes en attente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FriendRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Friends]
      summary: Envoyer une demande d'ami
      description: Envoie une demande à un utilisateur et le notifie (`friend:request`).
      operationId: sendFriendRequest
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendFriendRequestBody'
      responses:
        '201':
          description: Demande créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequest'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Relation déjà existante (amis ou demande en attente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Demande à soi-même
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /friends/requests/{id}/accept:
    post:
      tags: [Friends]
      summary: Accepter une demande d'ami
      description: Accepte une demande reçue et notifie le demandeur (`friend:accepted`).
      operationId: acceptFriendRequest
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Demande acceptée
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Demande déjà traitée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /friends/requests/{id}/decline:
    post:
      tags: [Friends]
      summary: Refuser une demande d'ami
      operationId: declineFriendRequest
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Demande refusée
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /friends/{userId}:
    delete:
      tags: [Friends]
      summary: Supprimer un ami
      operationId: removeFriend
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Ami supprimé
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'